+++ UI/test/com/kotcrab/vis/ui/test/manual/TestLauncher.java
@@ -19,11 +19,11 @@ package com.kotcrab.vis.ui.test.manual;
 import com.badlogic.gdx.Application.ApplicationType;
 import com.badlogic.gdx.ApplicationAdapter;
 import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.Input.Keys;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
 import com.badlogic.gdx.graphics.GL20;
-import com.badlogic.gdx.scenes.scene2d.Actor;
-import com.badlogic.gdx.scenes.scene2d.Stage;
+import com.badlogic.gdx.scenes.scene2d.*;
 import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 import com.badlogic.gdx.utils.viewport.ScreenViewport;
@@ -58,7 +58,7 @@ class TestApplication extends ApplicationAdapter {
 		VisUI.load(SkinScale.X1);
 
 		stage = new Stage(new ScreenViewport());
-		Table root = new Table();
+		final Table root = new Table();
 		root.setFillParent(true);
 		stage.addActor(root);
 
@@ -83,6 +83,28 @@ class TestApplication extends ApplicationAdapter {
 		stage.addActor(new TestValidator());
 		stage.addActor(new TestBuilders());
 //		stage.addActor(new TestTabbedPane());
+//		stage.addActor(new TestFlowGroup());
+
+		stage.addListener(new InputListener() {
+			boolean debug = false;
+
+			@Override
+			public boolean keyDown (InputEvent event, int keycode) {
+				if (keycode == Keys.F12) {
+					debug = !debug;
+					root.setDebug(debug, true);
+					for (Actor actor : stage.getActors()) {
+						if (actor instanceof Group) {
+							Group group = (Group) actor;
+							group.setDebug(debug, true);
+						}
+					}
+					return true;
+				}
+
+				return false;
+			}
+		});
 	}
 
 	private void createMenus () {